link: split out platform specific code
Move platform specific code into separate files. No other changes
intended.
